Adds two more extraction methods to Mods Extractor.
